+++ noPhone/View/recodeView.swift
@@ -16,56 +16,8 @@ struct RecodeView: View {
                                 .foregroundColor(.gray)
                                 .padding()
                         } else {
-                            List{
-                                if graph {
-                                    withAnimation {
-                                        LineView(recodeTimeList: recodeTimeList)
-                                            .transition(.move(edge: .bottom).combined(with: .opacity)) // 下からフェードイン
-                                    }
-                                }
-                                
-                                ForEach(recodeTimeList.indices, id: \.self) { index in
-                                    Section{
-                                        VStack{
-                                            HStack{
-                                                Spacer()
-                                                Text(formattedDate(date: recodeTimeList[index].date))
-                                                    .font(.title3)
-                                                
-                                                Spacer()
-                                                if recodeTimeList[index].close {
-                                                    Text("Success")
-                                                        .font(.title2)
-                                                        .foregroundColor(.blue)
-                                                } else {
-                                                    Text("Failed")
-                                                        .font(.title)
-                                                        .foregroundColor(.red)
-                                                }
-                                                Spacer()
-                                                
-                                            }
-                                            HStack {
-                                                Image(systemName: "timer") // アイコン
-                                                    .resizable()
-                                                    .frame(width: 20, height: 20)
-                                                RecoTimer(timer: recodeTimeList[index].settingtime, close: recodeTimeList[index].close,on:false)
-                                                Spacer()
-                                                Image(systemName: "person.badge.clock") // アイコン
-                                                    .resizable()
-                                                    .frame(width: 20, height: 20)
-                                                
-                                                RecoTimer(timer: recodeTimeList[index].realtime, close: recodeTimeList[index].close,on:true)
-                                                Spacer()
-                                            }
-                                        }.padding(.vertical, 8)
-                                            .listRowBackground(recodeTimeList[index].close == true ? ButtonColor : lightPink)
-                                    }
-                                }
-                            }
-                            .scrollContentBackground(.hidden)
-                            .background(GradientBackgroundView())
-                            .scrollIndicators(.hidden)
+                            RecoListView(recodeTimeList: recodeTimeList, graph: $graph)
+                            
                         }
                     }
                 
@@ -101,12 +53,6 @@ struct RecodeView: View {
             return Array(decoded.reversed()) // ここで順番を逆にする
         }
         return []
-    }
-    
-    func formattedDate(date: Date) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd" // カスタムフォーマット
-        return formatter.string(from: date)
     }
 }
